Add openapi + swagger ui
