+++ nym-network-monitor/Cargo.toml
@@ -22,8 +22,10 @@ dashmap = "5"
 serde_json = "1"
 serde = { version = "1", features = ["derive"] }
 petgraph = "0.6.5"
-log = { version = "0.4" }
-
+log = { workspace = true }
+clap = { workspace = true }
+utoipa = { workspace = true, features = ["axum_extras"] }
+utoipa-swagger-ui = { workspace = true, features = ["axum"] }
 
 # internal
 nym-sdk = { path = "../sdk/rust/nym-sdk" }
