+++ nym-network-monitor/locustfile.py
@@ -1,6 +1,7 @@
 from locust import HttpUser, task
 
+
 class SendMsg(HttpUser):
     @task
     def hello_world(self):
-        self.client.get("/")
+        self.client.post("/")
